+++ include/ydsh/ydsh.h
@@ -240,6 +240,7 @@ int DSState_eval(DSState *st, const char *sourceName, const char *data, unsigned
  * open file and evaluate. if e is not null, set error info.
  * set SCRIPT_DIR to dirname of fileName.
  * before evaluation reset line number.
+ * equivalent to DSState_loadModule(DS_MOD_FULLPATH | DS_MOD_SEPARATE_CTX)
  * @param st
  * @param fileName
  * not null
@@ -257,9 +258,20 @@ int DSState_loadAndEval(DSState *st, const char *fileName, DSError *e);
 
 /* for module loading option */
 /**
+ * load module as fullpath. so not allow cascading module search
+ * (not search from LOCAL_MOD_DIR and SYSTEM_MOD_DIR)
+ */
+#define DS_MOD_FULLPATH      ((unsigned int) (1u << 0u))
+
+/**
  * ignore ENOENT error
  */
-#define DS_MOD_IGNORE_ENOENT ((unsigned int) (1u << 0u))
+#define DS_MOD_IGNORE_ENOENT ((unsigned int) (1u << 1u))
+
+/**
+ * evaluate module in separate module context
+ */
+#define DS_MOD_SEPARATE_CTX  ((unsigned int) (1u << 2u))
 
 /**
  * open file as module. if e is not null, set error info.
@@ -267,7 +279,7 @@ int DSState_loadAndEval(DSState *st, const char *fileName, DSError *e);
  * @param st
  * not null.
  * @param fileName
- * not null. if realtive path, search from module loading path
+ * not null
  * @param option
  * @param e
  * may be null