+++ test/extra/extra_test.cpp
@@ -186,9 +186,18 @@ struct APITest : public ExpectOutput {
 
     ~APITest() override {
         DSState_delete(&this->state);
-        DSState_delete(&this->state);
     }
 };
+
+TEST_F(APITest, modFullpath) {
+    DSError e;
+    int r = DSState_loadModule(this->state, "edit", DS_MOD_FULLPATH, &e);   // not load 'edit'
+    ASSERT_EQ(1, r);
+    ASSERT_EQ(DS_ERROR_KIND_TYPE_ERROR, e.kind);
+    ASSERT_STREQ("NotFoundMod", e.name);
+    ASSERT_EQ(0, e.lineNum);
+    DSError_release(&e);
+}
 
 TEST_F(APITest, mod) {
     DSError e;